Issue-#14: - Sort methods in Utils. - Move fraction a handling method and a constant to Fraction class. - Add some JavaDoc to trivial method. - Improve a handling of input stream when a file is loaded.
